refactor: split dashboard tab views
