+++ src/views/SettingsView.tsx
@@ -1,0 +1,384 @@
+import { Match, Show, Switch } from "solid-js";
+
+import { formatBytes, formatRelativeTime, isTauriRuntime } from "../app/utils";
+
+import Button from "../components/Button";
+import { HardDrive, RefreshCcw, Shield, Smartphone } from "lucide-solid";
+
+export type SettingsViewProps = {
+  mode: "host" | "client" | null;
+  baseUrl: string;
+  headerStatus: string;
+  busy: boolean;
+  developerMode: boolean;
+  toggleDeveloperMode: () => void;
+  stopHost: () => void;
+  engineSource: "path" | "sidecar";
+  setEngineSource: (value: "path" | "sidecar") => void;
+  isWindows: boolean;
+  defaultModelLabel: string;
+  defaultModelRef: string;
+  openDefaultModelPicker: () => void;
+  showThinking: boolean;
+  toggleShowThinking: () => void;
+  modelVariantLabel: string;
+  editModelVariant: () => void;
+  updateAutoCheck: boolean;
+  toggleUpdateAutoCheck: () => void;
+  updateStatus: {
+    state: string;
+    lastCheckedAt?: number | null;
+    version?: string;
+    date?: string;
+    notes?: string;
+    totalBytes?: number | null;
+    downloadedBytes?: number;
+    message?: string;
+  } | null;
+  updateEnv: { supported?: boolean; reason?: string | null } | null;
+  appVersion: string | null;
+  checkForUpdates: () => void;
+  downloadUpdate: () => void;
+  installUpdateAndRestart: () => void;
+  anyActiveRuns: boolean;
+  onResetStartupPreference: () => void;
+  openResetModal: (mode: "onboarding" | "all") => void;
+  resetModalBusy: boolean;
+  pendingPermissions: unknown;
+  events: unknown;
+  safeStringify: (value: unknown) => string;
+};
+
+export default function SettingsView(props: SettingsViewProps) {
+  const updateState = () => props.updateStatus?.state ?? "idle";
+  const updateNotes = () => props.updateStatus?.notes ?? null;
+  const updateVersion = () => props.updateStatus?.version ?? null;
+  const updateDate = () => props.updateStatus?.date ?? null;
+  const updateLastCheckedAt = () => props.updateStatus?.lastCheckedAt ?? null;
+  const updateDownloadedBytes = () => props.updateStatus?.downloadedBytes ?? null;
+  const updateTotalBytes = () => props.updateStatus?.totalBytes ?? null;
+  const updateErrorMessage = () => props.updateStatus?.message ?? null;
+
+  return (
+    <section class="space-y-6">
+      <div class="bg-zinc-900/30 border border-zinc-800/50 rounded-2xl p-5 space-y-3">
+        <div class="text-sm font-medium text-white">Connection</div>
+        <div class="text-xs text-zinc-500">{props.headerStatus}</div>
+        <div class="text-xs text-zinc-600 font-mono">{props.baseUrl}</div>
+        <div class="pt-2 flex flex-wrap gap-2">
+          <Button variant="secondary" onClick={props.toggleDeveloperMode}>
+            <Shield size={16} />
+            {props.developerMode ? "Disable Developer Mode" : "Enable Developer Mode"}
+          </Button>
+          <Show when={props.mode === "host"}>
+            <Button variant="danger" onClick={props.stopHost} disabled={props.busy}>
+              Stop engine
+            </Button>
+          </Show>
+          <Show when={props.mode === "client"}>
+            <Button variant="outline" onClick={props.stopHost} disabled={props.busy}>
+              Disconnect
+            </Button>
+          </Show>
+        </div>
+
+        <Show when={isTauriRuntime() && props.mode === "host"}>
+          <div class="pt-4 border-t border-zinc-800/60 space-y-3">
+            <div class="text-xs text-zinc-500">Engine source</div>
+            <div class="grid grid-cols-2 gap-2">
+              <Button
+                variant={props.engineSource === "path" ? "secondary" : "outline"}
+                onClick={() => props.setEngineSource("path")}
+                disabled={props.busy}
+              >
+                PATH
+              </Button>
+              <Button
+                variant={props.engineSource === "sidecar" ? "secondary" : "outline"}
+                onClick={() => props.setEngineSource("sidecar")}
+                disabled={props.busy || props.isWindows}
+                title={props.isWindows ? "Sidecar is not supported on Windows yet" : ""}
+              >
+                Sidecar
+              </Button>
+            </div>
+            <div class="text-[11px] text-zinc-600">
+              PATH uses your installed OpenCode (default). Sidecar will use a bundled binary when available.
+              <Show when={props.isWindows}>
+                <span class="text-zinc-500"> Sidecar is currently unavailable on Windows.</span>
+              </Show>
+            </div>
+          </div>
+        </Show>
+      </div>
+
+      <div class="bg-zinc-900/30 border border-zinc-800/50 rounded-2xl p-5 space-y-4">
+        <div>
+          <div class="text-sm font-medium text-white">Model</div>
+          <div class="text-xs text-zinc-500">Defaults + thinking controls for runs.</div>
+        </div>
+
+        <div class="flex items-center justify-between bg-zinc-950 p-3 rounded-xl border border-zinc-800 gap-3">
+          <div class="min-w-0">
+            <div class="text-sm text-zinc-200 truncate">{props.defaultModelLabel}</div>
+            <div class="text-xs text-zinc-600 font-mono truncate">{props.defaultModelRef}</div>
+          </div>
+          <Button
+            variant="outline"
+            class="text-xs h-8 py-0 px-3 shrink-0"
+            onClick={props.openDefaultModelPicker}
+            disabled={props.busy}
+          >
+            Change
+          </Button>
+        </div>
+
+        <div class="flex items-center justify-between bg-zinc-950 p-3 rounded-xl border border-zinc-800 gap-3">
+          <div class="min-w-0">
+            <div class="text-sm text-zinc-200">Thinking</div>
+            <div class="text-xs text-zinc-600">Show thinking parts (Developer mode only).</div>
+          </div>
+          <Button
+            variant="outline"
+            class="text-xs h-8 py-0 px-3 shrink-0"
+            onClick={props.toggleShowThinking}
+            disabled={props.busy}
+          >
+            {props.showThinking ? "On" : "Off"}
+          </Button>
+        </div>
+
+        <div class="flex items-center justify-between bg-zinc-950 p-3 rounded-xl border border-zinc-800 gap-3">
+          <div class="min-w-0">
+            <div class="text-sm text-zinc-200">Model variant</div>
+            <div class="text-xs text-zinc-600 font-mono truncate">{props.modelVariantLabel}</div>
+          </div>
+          <Button
+            variant="outline"
+            class="text-xs h-8 py-0 px-3 shrink-0"
+            onClick={props.editModelVariant}
+            disabled={props.busy}
+          >
+            Edit
+          </Button>
+        </div>
+      </div>
+
+      <div class="bg-zinc-900/30 border border-zinc-800/50 rounded-2xl p-5 space-y-3">
+        <div class="flex items-start justify-between gap-4">
+          <div>
+            <div class="text-sm font-medium text-white">Updates</div>
+            <div class="text-xs text-zinc-500">Keep OpenWork up to date.</div>
+          </div>
+          <div class="text-xs text-zinc-600 font-mono">{props.appVersion ? `v${props.appVersion}` : ""}</div>
+        </div>
+
+        <Show
+          when={!isTauriRuntime()}
+          fallback={
+            <Show
+              when={props.updateEnv && props.updateEnv.supported === false}
+              fallback={
+                <>
+                  <div class="flex items-center justify-between bg-zinc-950 p-3 rounded-xl border border-zinc-800">
+                    <div class="space-y-0.5">
+                      <div class="text-sm text-white">Automatic checks</div>
+                      <div class="text-xs text-zinc-600">Once per day (quiet)</div>
+                    </div>
+                    <button
+                      class={`px-3 py-1 rounded-full text-xs font-medium border transition-colors ${
+                        props.updateAutoCheck
+                          ? "bg-white/10 text-white border-white/20"
+                          : "text-zinc-500 border-zinc-800 hover:text-white"
+                      }`}
+                      onClick={props.toggleUpdateAutoCheck}
+                    >
+                      {props.updateAutoCheck ? "On" : "Off"}
+                    </button>
+                  </div>
+
+                  <div class="flex items-center justify-between gap-3 bg-zinc-950 p-3 rounded-xl border border-zinc-800">
+                    <div class="space-y-0.5">
+                      <div class="text-sm text-white">
+                        <Switch>
+                          <Match when={updateState() === "checking"}>Checking...</Match>
+                          <Match when={updateState() === "available"}>Update available: v{updateVersion()}</Match>
+                          <Match when={updateState() === "downloading"}>Downloading...</Match>
+                          <Match when={updateState() === "ready"}>Ready to install: v{updateVersion()}</Match>
+                          <Match when={updateState() === "error"}>Update check failed</Match>
+                          <Match when={true}>Up to date</Match>
+                        </Switch>
+                      </div>
+                      <Show when={updateState() === "idle" && updateLastCheckedAt()}>
+                        <div class="text-xs text-zinc-600">
+                          Last checked {formatRelativeTime(updateLastCheckedAt() as number)}
+                        </div>
+                      </Show>
+                      <Show when={updateState() === "available" && updateDate()}>
+                        <div class="text-xs text-zinc-600">Published {updateDate()}</div>
+                      </Show>
+                      <Show when={updateState() === "downloading"}>
+                        <div class="text-xs text-zinc-600">
+                          {formatBytes((updateDownloadedBytes() as number) ?? 0)}
+                          <Show when={updateTotalBytes() != null}>
+                            {` / ${formatBytes(updateTotalBytes() as number)}`}
+                          </Show>
+                        </div>
+                      </Show>
+                      <Show when={updateState() === "error"}>
+                        <div class="text-xs text-red-300">{updateErrorMessage()}</div>
+                      </Show>
+                    </div>
+
+                    <div class="flex items-center gap-2">
+                      <Button
+                        variant="outline"
+                        class="text-xs h-8 py-0 px-3"
+                        onClick={props.checkForUpdates}
+                        disabled={props.busy || updateState() === "checking" || updateState() === "downloading"}
+                      >
+                        Check
+                      </Button>
+
+                      <Show when={updateState() === "available"}>
+                        <Button
+                          variant="secondary"
+                          class="text-xs h-8 py-0 px-3"
+                          onClick={props.downloadUpdate}
+                          disabled={props.busy || updateState() === "downloading"}
+                        >
+                          Download
+                        </Button>
+                      </Show>
+
+                      <Show when={updateState() === "ready"}>
+                        <Button
+                          variant="secondary"
+                          class="text-xs h-8 py-0 px-3"
+                          onClick={props.installUpdateAndRestart}
+                          disabled={props.busy || props.anyActiveRuns}
+                          title={props.anyActiveRuns ? "Stop active runs to update" : ""}
+                        >
+                          Install & Restart
+                        </Button>
+                      </Show>
+                    </div>
+                  </div>
+
+                  <Show when={updateState() === "available" && updateNotes()}>
+                    <div class="rounded-xl bg-black/20 border border-zinc-800 p-3 text-xs text-zinc-400 whitespace-pre-wrap max-h-40 overflow-auto">
+                      {updateNotes()}
+                    </div>
+                  </Show>
+                </>
+              }
+            >
+              <div class="rounded-xl bg-black/20 border border-zinc-800 p-3 text-sm text-zinc-400">
+                {props.updateEnv?.reason ?? "Updates are not supported in this environment."}
+              </div>
+            </Show>
+          }
+        >
+          <div class="rounded-xl bg-black/20 border border-zinc-800 p-3 text-sm text-zinc-400">
+            Updates are only available in the desktop app.
+          </div>
+        </Show>
+      </div>
+
+      <div class="bg-zinc-900/30 border border-zinc-800/50 rounded-2xl p-5 space-y-3">
+        <div class="text-sm font-medium text-white">Startup</div>
+
+        <div class="flex items-center justify-between bg-zinc-950 p-3 rounded-xl border border-zinc-800">
+          <div class="flex items-center gap-3">
+            <div
+              class={`p-2 rounded-lg ${
+                props.mode === "host" ? "bg-indigo-500/10 text-indigo-400" : "bg-emerald-500/10 text-emerald-400"
+              }`}
+            >
+              <Show when={props.mode === "host"} fallback={<Smartphone size={18} />}>
+                <HardDrive size={18} />
+              </Show>
+            </div>
+            <span class="capitalize text-sm font-medium text-white">{props.mode} mode</span>
+          </div>
+          <Button variant="outline" class="text-xs h-8 py-0 px-3" onClick={props.stopHost} disabled={props.busy}>
+            Switch
+          </Button>
+        </div>
+
+        <Button variant="secondary" class="w-full justify-between group" onClick={props.onResetStartupPreference}>
+          <span class="text-zinc-300">Reset default startup mode</span>
+          <RefreshCcw size={14} class="text-zinc-500 group-hover:rotate-180 transition-transform" />
+        </Button>
+
+        <p class="text-xs text-zinc-600">
+          This clears your saved preference and shows mode selection on next launch.
+        </p>
+      </div>
+
+      <div class="bg-zinc-900/30 border border-zinc-800/50 rounded-2xl p-5 space-y-4">
+        <div>
+          <div class="text-sm font-medium text-white">Advanced</div>
+          <div class="text-xs text-zinc-500">Reset OpenWork local state to retest onboarding.</div>
+        </div>
+
+        <div class="flex items-center justify-between bg-zinc-950 p-3 rounded-xl border border-zinc-800 gap-3">
+          <div class="min-w-0">
+            <div class="text-sm text-zinc-200">Reset onboarding</div>
+            <div class="text-xs text-zinc-600">Clears OpenWork preferences and restarts the app.</div>
+          </div>
+          <Button
+            variant="outline"
+            class="text-xs h-8 py-0 px-3 shrink-0"
+            onClick={() => props.openResetModal("onboarding")}
+            disabled={props.busy || props.resetModalBusy || props.anyActiveRuns}
+            title={props.anyActiveRuns ? "Stop active runs to reset" : ""}
+          >
+            Reset
+          </Button>
+        </div>
+
+        <div class="flex items-center justify-between bg-zinc-950 p-3 rounded-xl border border-zinc-800 gap-3">
+          <div class="min-w-0">
+            <div class="text-sm text-zinc-200">Reset app data</div>
+            <div class="text-xs text-zinc-600">More aggressive. Clears OpenWork cache + app data.</div>
+          </div>
+          <Button
+            variant="danger"
+            class="text-xs h-8 py-0 px-3 shrink-0"
+            onClick={() => props.openResetModal("all")}
+            disabled={props.busy || props.resetModalBusy || props.anyActiveRuns}
+            title={props.anyActiveRuns ? "Stop active runs to reset" : ""}
+          >
+            Reset
+          </Button>
+        </div>
+
+        <div class="text-xs text-zinc-600">
+          Requires typing <span class="font-mono text-zinc-400">RESET</span> and will restart the app.
+        </div>
+      </div>
+
+      <Show when={props.developerMode}>
+        <section>
+          <h3 class="text-sm font-medium text-zinc-400 uppercase tracking-wider mb-4">Developer</h3>
+
+          <div class="grid md:grid-cols-2 gap-4">
+            <div class="bg-zinc-900/30 border border-zinc-800/50 rounded-2xl p-4">
+              <div class="text-xs text-zinc-500 mb-2">Pending permissions</div>
+              <pre class="text-xs text-zinc-200 whitespace-pre-wrap break-words max-h-64 overflow-auto">
+                {props.safeStringify(props.pendingPermissions)}
+              </pre>
+            </div>
+            <div class="bg-zinc-900/30 border border-zinc-800/50 rounded-2xl p-4">
+              <div class="text-xs text-zinc-500 mb-2">Recent events</div>
+              <pre class="text-xs text-zinc-200 whitespace-pre-wrap break-words max-h-64 overflow-auto">
+                {props.safeStringify(props.events)}
+              </pre>
+            </div>
+          </div>
+        </section>
+      </Show>
+    </section>
+  );
+}
